Multiple object meshes combined to one huge mesh

+++ Particle.h
@@ -5,23 +5,23 @@
 
 class Particle {
 private:
-    Eigen::Vector3f m_position; // Position of the particle
-    Eigen::Vector3f m_velocity; // Velocity of the particle
-    Eigen::Quaternionf m_orientation; // Orientation of the particle (stored as a quaternion)
+    Eigen::Vector3d m_position; // Position of the particle
+    Eigen::Vector3d m_velocity; // Velocity of the particle
+    Eigen::Quaterniond m_orientation; // Orientation of the particle (stored as a quaternion)
     
 
 public:
     Particle() : m_position(0, 0, 0), m_velocity(0, 0, 0), m_orientation(1, 0, 0, 0) {}
 
     // Setter functions for position, velocity, and orientation
-    void setPosition(const Eigen::Vector3f& position) { m_position = position; }
-    void setVelocity(const Eigen::Vector3f& velocity) { m_velocity = velocity; }
-    void setOrientation(const Eigen::Quaternionf& orientation) { m_orientation = orientation.normalized(); }
+    void setPosition(const Eigen::Vector3d& position) { m_position = position; }
+    void setVelocity(const Eigen::Vector3d& velocity) { m_velocity = velocity; }
+    void setOrientation(const Eigen::Quaterniond& orientation) { m_orientation = orientation.normalized(); }
 
     // Getter functions for position, velocity, and orientation
-    Eigen::Vector3f getPosition() const { return m_position; }
-    Eigen::Vector3f getVelocity() const { return m_velocity; }
-    Eigen::Quaternionf getOrientation() const { return m_orientation; }
+    Eigen::Vector3d getPosition() const { return m_position; }
+    Eigen::Vector3d getVelocity() const { return m_velocity; }
+    Eigen::Quaterniond getOrientation() const { return m_orientation; }
 
     // Other functions for particle behavior or manipulation...
 };
